refactor: plast-mem bug fix and Haru refactor (#29)
* refacor: fix plast-mem bugs and Haru message window

* fix: multiple plast-mem and Haru bugs

* doc: update

* Update examples/haru/src/hooks/use-haru.ts

* fix: gemini review comments

+++ crates/core/src/memory/episodic/mod.rs
@@ -154,7 +154,7 @@ impl EpisodicMemory {
       SELECT id, 1.0 / (60 + r) AS s FROM semantic
     ),
     rrf_score AS (
-      SELECT id, SUM(s) AS score
+      SELECT id, SUM(s)::float8 AS score
       FROM rrf
       GROUP BY id
     )
