+++ crates/core/src/memory/semantic/mod.rs
@@ -106,7 +106,7 @@ impl SemanticMemory {
       SELECT id, 1.0 / (60 + r) AS s FROM semantic
     ),
     rrf_score AS (
-      SELECT id, SUM(s) AS score
+      SELECT id, SUM(s)::float8 AS score
       FROM rrf
       GROUP BY id
     )
